feat(cv_3d): fundamental

+++ cv-3d/src/block_match/mod.rs
@@ -3,9 +3,9 @@ use nalgebra::{DMatrix, Vector2};
 use crate::frame::Frame;
 
 pub trait BlockMatch {
-    fn match_block(&self, pt1: &Vector2<u32>, pt2: &Vector2<u32>) -> f32;
-    fn better(&self, last_score: f32, current_score: f32) -> bool;
-    fn best(&self, score: f32) -> bool;
+    fn match_block(&self, pt1: &Vector2<u32>, pt2: &Vector2<u32>) -> f64;
+    fn better(&self, last_score: f64, current_score: f64) -> bool;
+    fn best(&self, score: f64) -> bool;
 }
 
 #[derive(Debug, Clone)]
@@ -13,14 +13,14 @@ pub struct Ncc<'a> {
     ncc_window_size: Option<u32>,
     image1: Frame<'a>,
     image2: Frame<'a>,
-    pass_number: f32,
+    pass_number: f64,
 }
 
 impl<'a> Ncc<'a> {
     pub fn new(
         image1: Frame<'a>,
         image2: Frame<'a>,
-        pass_number: f32,
+        pass_number: f64,
         ncc_window_size: Option<u32>,
     ) -> Self {
         Self {
@@ -33,7 +33,7 @@ impl<'a> Ncc<'a> {
 }
 
 impl BlockMatch for Ncc<'_> {
-    fn match_block(&self, pt1: &Vector2<u32>, pt2: &Vector2<u32>) -> f32 {
+    fn match_block(&self, pt1: &Vector2<u32>, pt2: &Vector2<u32>) -> f64 {
         let ncc_window_size = self.ncc_window_size.unwrap_or(3);
         let block1 = self.image1.data.slice(
             (
@@ -66,14 +66,14 @@ impl BlockMatch for Ncc<'_> {
             .sqrt();
         numerator / (divisor + 1e-18)
     }
-    fn better(&self, last: f32, current: f32) -> bool {
+    fn better(&self, last: f64, current: f64) -> bool {
         if current < last {
             true
         } else {
             false
         }
     }
-    fn best(&self, score: f32) -> bool {
+    fn best(&self, score: f64) -> bool {
         if score < self.pass_number {
             true
         } else {
