+++ cv-3d/src/lib.rs
@@ -14,12 +14,12 @@ pub mod fundamental;
 pub mod triangluate;
 pub mod utils;
 pub trait ToNaVector2 {
-    fn to_vector2(&self) -> Vector2<f32>;
+    fn to_vector2(&self) -> Vector2<f64>;
 }
 
 impl ToNaVector2 for KeyPoint {
-    fn to_vector2(&self) -> Vector2<f32> {
-        Vector2::new(self.pt.x, self.pt.y)
+    fn to_vector2(&self) -> Vector2<f64> {
+        Vector2::new(self.pt.x as f64, self.pt.y as f64)
     }
 }
